getSugangSnuSearchContent 및 getPageCount를 SugangSnuFetchService로 이동
